Split integration tests into modules, one per command
