+++ tests/integration_tests.rs
@@ -1,0 +1,1 @@
+mod integration;
